third_party/WebKit/LayoutTests/fast/scrolling/fractional-scroll-offset-document-expected.txt

Verifies that document scrolling supports fractional offset.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS document.documentElement.scrollTop is within 0.00001 of 0
PASS document.documentElement.scrollLeft is within 0.00001 of 0
PASS document.documentElement.scrollTop is within 0.00001 of 10.2
PASS document.documentElement.scrollLeft is within 0.00001 of 20.5
PASS document.documentElement.scrollTop is within 0.00001 of 31.1
PASS document.documentElement.scrollLeft is within 0.00001 of 9.6
PASS document.documentElement.scrollTop is within 0.00001 of 0
PASS document.documentElement.scrollLeft is within 0.00001 of 0
